Added channel and layer to brain stitcher

+++ src/pipeline/scripts/create_precomputed_from_volume.py
@@ -19,6 +19,8 @@ sys.path.append(PIPELINE_ROOT.as_posix())
 from library.image_manipulation.filelocation_manager import FileLocationManager
 from library.image_manipulation.neuroglancer_manager import NumpyToNeuroglancer
 from library.utilities.utilities_process import get_hostname, read_image
+from library.controller.sql_controller import SqlController
+
 from library.utilities.utilities_mask import normalize16
 
 
@@ -26,7 +28,11 @@ def create_precomputed(animal, volume_file, um):
     chunk = 64
     chunks = (chunk, chunk, chunk)
     fileLocationManager = FileLocationManager(animal)
-    INPUT = os.path.join(fileLocationManager.prep, 'CH1')
+    sqlController = SqlController(animal)
+    fileLocationManager = FileLocationManager(animal)
+    xy = sqlController.scan_run.resolution * 1000
+    z = sqlController.scan_run.zresolution * 1000
+    INPUT = os.path.join(fileLocationManager.prep, 'C1', 'registration')
     volumepath = os.path.join(INPUT, volume_file)
     if not os.path.exists(volumepath):
         print(f'{volumepath} does not exist, exiting.')
@@ -36,12 +42,16 @@ def create_precomputed(animal, volume_file, um):
     outpath = outpath.split('.')[0]
     ext = outpath.split('.')[0]
     IMAGE_OUTPUT = os.path.join(fileLocationManager.neuroglancer_data, f'{outpath}')
-    scale = um * 1000
-    scales = (scale, scale, scale)
-    if 'godzilla' in get_hostname():
+    #scale = um * 1000
+    #scales = (scale, scale, scale)
+
+    xy *=  10
+    scales = (int(xy), int(xy), int(z))
+    print(f'scales={scales}')
+
+    if 'mothra' in get_hostname() and os.path.exists(IMAGE_OUTPUT):
         print(f'Cleaning {IMAGE_OUTPUT}')
-        if os.path.exists(IMAGE_OUTPUT):
-            shutil.rmtree(IMAGE_OUTPUT)
+        shutil.rmtree(IMAGE_OUTPUT)
 
 
     os.makedirs(IMAGE_OUTPUT, exist_ok=True)
@@ -50,11 +60,10 @@ def create_precomputed(animal, volume_file, um):
         volume, _ = nrrd.read(volume_file)
     else:
         volume = read_image(volumepath)
-    #volume = np.swapaxes(volume, 0, 2)
+    volume = np.swapaxes(volume, 0, 2)
     num_channels = 1
     volume_size = volume.shape
-    print(f'volume shape={volume.shape} dtype={volume.dtype}')
-    volume = normalize16(volume)
+    #volume = normalize16(volume)
     print(f'volume shape={volume.shape} dtype={volume.dtype}')
 
     ng = NumpyToNeuroglancer(
@@ -82,7 +91,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Work on Animal')
     parser.add_argument('--animal', help='Enter the animal', required=True)
     parser.add_argument('--volume', help='Enter the name of the volume file', required=False, default='result.tif')
-    parser.add_argument('--um', help="size of Allen atlas in micrometers", required=False, default=10, type=int)
+    parser.add_argument('--um', help="size of volume in micrometers", required=False, default=10, type=int)
     args = parser.parse_args()
     animal = args.animal
     volume = args.volume
